+++ com.unity.render-pipelines.high-definition/Editor/Lighting/Reflection/CameraSettingsUI.Drawers.cs
@@ -83,6 +83,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             }
 
             PropertyFieldWithFlagToggleIfDisplayed(CameraSettingsFields.flipYMode, serialized.flipYMode, EditorGUIUtility.TrTextContent("Flip Y"), @override.camera, displayedFields.camera, overridableFields.camera);
+            PropertyFieldWithFlagToggleIfDisplayed(CameraSettingsFields.probeLayerMask, serialized.probeLayerMask, EditorGUIUtility.TrTextContent("Probe Layer Mask"), @override.camera, displayedFields.camera, overridableFields.camera);
             PropertyFieldWithFlagToggleIfDisplayed(CameraSettingsFields.customRenderingSettings, serialized.customRenderingSettings, EditorGUIUtility.TrTextContent("Custom Frame Settings"), @override.camera, displayedFields.camera, overridableFields.camera);
 
             if ((displayedFields.camera & CameraSettingsFields.frameSettings) != 0)